♻️ move qc functions to submodules of qc
